+++ src/features/XIT/FIN/FINPR.vue
@@ -80,7 +80,7 @@ function profitClass(value: number) {
 </template>
 
 <style scoped>
-table tr > *:not(:first-child) {
+table tr > :not(:first-child) {
   text-align: right;
 }
 </style>