+++ src/features/XIT/FINBS/FINBS.vue
@@ -310,7 +310,7 @@ const sections = [
 </template>
 
 <style scoped>
-table tr > *:not(:first-child) {
+table tr > :not(:first-child) {
   text-align: right;
 }
 </style>