+++ src/infrastructure/prun-api/data/tiles.types.d.ts
@@ -1,12 +1,12 @@
 declare namespace PrunApi {
-  export interface Tile {
+  interface Tile {
     id: string;
     parentId: null | string;
     content: null | string;
     container: TileContainer | null;
   }
 
-  export interface TileContainer {
+  interface TileContainer {
     child1Id: string;
     child2Id: string;
     dividerPosition: number;